Update button edit produksi, projek, projek-rnd

+++ resources/views/livewire/projek-table.blade.php
@@ -79,13 +79,7 @@
                                 <div class="text-slate-800 dark:text-slate-100">{{ $projeks->firstItem() + $index }}</div>
                             </td>
                             <td class="px-6 py-3">
-                                @if ($projek->status !== 'Konfirmasi')
-                                @can('edit-projek')
-                                    <strong><u><a href="{{ route('projeks.edit', $projek->id) }}">{{ $projek->kode_projek }}</a></u></strong>
-                                @endcan
-                                @else
-                                    <span class="text-gray-500">{{ $projek->kode_projek }}</span>
-                                @endif
+                                <strong>{{ $projek->kode_projek }}</strong>
                             </td>
                             <td class="px-6 py-3">{{ $projek->mulai_projek }}</td>
                             <td class="px-6 py-3">{{ $projek->selesai_projek }}</td>
@@ -95,19 +89,30 @@
                             {{-- <td class="px-6 py-3">Rp {{ number_format($projek->produksiDetails->sum('sub_total'), 2, ',', '.') }}</td> --}}
                             <td class="px-6 py-4">
                                 <div class="row flex space-x-2">
+                                    @if($projek->status !== 'Konfirmasi')
+                                        @can('edit-projek')
+                                            <a href="{{ route('projeks.edit', $projek->id) }}" class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-yellow-600 hover:border-yellow-600 focus:text-white focus:bg-yellow-600 focus:border-yellow-600 active:border-yellow-600 active:text-white active:bg-yellow-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none">
+                                                <svg class="w-[16px] h-[16px] text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
+                                                <path stroke="currentColor" stroke-width="2" d="m14.304 4.844 2.852 2.852M7 7H4a1 1 0 0 0-1 1v10a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-4.5m2.409-9.91a2.017 2.017 0 0 1 0 2.853l-6.844 6.844L8 14l.713-3.565 6.844-6.844a2.015 2.015 0 0 1 2.852 0Z"/>
+                                                </svg>
+                                            </a>
+                                        @endcan
+                                    @endif
                                     @if ($projek->status === 'Konfirmasi')
-                                        <button wire:click="deleteProjeks({{ $projek->id }})"
-                                            data-modal-target="deleteproduksi-modal" data-modal-toggle="deleteproduksi-modal"
-                                            class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-red-600 hover:border-red-600 focus:text-white focus:bg-red-600 focus:border-red-600 active:border-red-600 active:text-white active:bg-red-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
-                                            type="button">
-                                            <svg class="w-[16px] h-[16px] text-gray-800 dark:text-white" aria-hidden="true"
-                                                xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none"
-                                                viewBox="0 0 24 24">
-                                                <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
-                                                    stroke-width="2"
-                                                    d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z" />
-                                            </svg>
-                                        </button>
+                                        @can('hapus-projek')
+                                            <button wire:click="deleteProjeks({{ $projek->id }})"
+                                                data-modal-target="deleteproduksi-modal" data-modal-toggle="deleteproduksi-modal"
+                                                class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-red-600 hover:border-red-600 focus:text-white focus:bg-red-600 focus:border-red-600 active:border-red-600 active:text-white active:bg-red-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+                                                type="button">
+                                                <svg class="w-[16px] h-[16px] text-gray-800 dark:text-white" aria-hidden="true"
+                                                    xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none"
+                                                    viewBox="0 0 24 24">
+                                                    <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
+                                                        stroke-width="2"
+                                                        d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z" />
+                                                </svg>
+                                            </button>
+                                        @endcan
                                     @endif
                                 </div>
                             </td>
